requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.12
idna==3.3
progressbar2==3.55.0
python-utils==3.1.0
requests==2.27.1
six==1.16.0
urllib3==1.26.8
